Convert step.options to snake_case
